requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
gast==0.2.2
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-pasta==0.1.4
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
h5py==2.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.17
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.7
numpy==1.16.2
opencv-python==4.1.0.25
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
pytesseract==0.3.0
python-dateutil==2.8.0
pytube==9.5.1
pytz==2018.9
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
rsa==4.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
tb-nightly==1.14.0a20190301
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.1
widgetsnbextension==3.4.2
wrapt==1.11.1
